플레이어 무적 구현 // 플레이어 - Boss B 충돌 처리 구현

+++ Project/Project/CWeapon.h
@@ -24,7 +24,7 @@ public:
 	int m_type;
 
 	int m_rounds;
-	float m_attack;
+	int m_attack;
 
 	float m_range;
 	float m_cooldown;
